refactor common service functions into one.
